% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietlein_data-data.R
\docType{data}
\name{dietlein_data}
\alias{dietlein_data}
\title{Response data from Dietlein et al}
\format{An object of class \code{"data.frame"}.}
\source{
\href{http://cancerdiscovery.aacrjournals.org/content/4/5/592/suppl/DC1}{Supplementary
  data table 1}
}
\usage{
data(dietlein_data)
}
\description{
Data from the paper 'A Functional Cancer Genomics Screen Identifies a
Druggable Synthetic Lethal Interaction between MSH3 and PRKDC' in Cancer
Discovery by Dietlein et al 2014.  67 cell lines were treated with the
DNA-PKcs (DNA-dependent protein kinase catalytic subunit) inhibitor KU60648
}
\references{
Dietlein et al. (2014)  Cancer Discovery May 1, 2014 4; 592
  (\href{http://www.ncbi.nlm.nih.gov/pubmed/24556366}{PubMed})
}
\keyword{datasets}

