% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShinyVisFunctions.R
\name{get_shiny_cell_lines}
\alias{get_shiny_cell_lines}
\title{Convert tissue to cell lines}
\usage{
get_shiny_cell_lines(con, tissue_id)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{cell_lines}{A vector of cell line identifiers}
}
\value{
A \code{vector} of cell lines for the provided tissue
}
\description{
This function creates a \code{vector} of cell lines from a tissue
}

