% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\name{getAffyData}
\alias{getAffyData}
\title{Extract affymetrix expression data}
\usage{
getAffyData(con, genes, cell_lines)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{genes}{A vector of gene symbols}

\item{cell_lines}{A vectore of cell line identifiers}
}
\value{
A \code{data.frame} containing the affymetrix gene expression data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} containing the affymetrix  gene expression data from the database for the requested cell lines and genes.
}

