% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\name{getDrugData}
\alias{getDrugData}
\title{Extract drug response data}
\usage{
getDrugData(con, drugs, cell_lines)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{drugs}{A vector of compound identifiers}

\item{cell_lines}{A vector of cell line identifiers}
}
\value{
A \code{data.frame} containing the drug response data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} containing the drug response data from the database.
}

