% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDrugData_custom.R
\name{getDrugData_custom}
\alias{getDrugData_custom}
\title{Extract custom drug response data}
\usage{
getDrugData_custom(df, drugs, cell_lines)
}
\arguments{
\item{df}{A \code{data.frame} object containing the data which should include columns names compound_id, unified_id, endpoint, original and value}

\item{drugs}{A vector of compound identifiers or compound/endpoint identifiers}

\item{cell_lines}{A vector of cell line identifiers}
}
\value{
A \code{data.frame} containing the drug response data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} containing the drug response data from a user provided data frame
}

