% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getHybcapData.R
\name{getHybcapData}
\alias{getHybcapData}
\title{Extract hybrid capture sequencing data}
\usage{
getHybcapData(con, genes, cell_lines)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{genes}{A vector of gene symbols}

\item{cell_lines}{A vector of cell line identifiers}
}
\value{
A \code{data.frame} containing the hybrid capture sequencing data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} containing the hybrid capture sequencing from the database for the requested cell lines and genes.  Gene/sample pairs where there exists at least one missense, nonsense or framshift mutation are categorised as 1, those that don't are categorised as 0.
}

