% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponseData.R
\name{getResponseData}
\alias{getResponseData}
\title{Generic function to get response data from multiple sources}
\usage{
getResponseData(src, drugs, cell_lines, resp_type = "ccle")
}
\arguments{
\item{src}{A \code{SQLiteConnection} object to the database (ccle) or \code{data.frame} object (custom)}

\item{drugs}{A vector of compound identifiers}

\item{cell_lines}{A vector of cell line identifiers}

\item{resp_type}{The type of response data: \code{ccle, custom}}
}
\value{
A \code{data.frame} containing the response data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} containing the response data from a variety of sourses.
}

