% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTissueInfo.R
\name{getTissueInfo}
\alias{getTissueInfo}
\title{Get tissue info for cell lines}
\usage{
getTissueInfo(con, tissue_info = "ccle")
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{tissue_info}{Source of tissue information: \code{ccle} uses the ccle sample information, \code{crude} ensures that all cell lines are returned but mixes data from different sources, \code{custom} can be one of any id_type present in the cell_line_ids table for extra control}
}
\value{
A \code{data.frame} containing tissue information for available cell lines
}
\description{
This function creates a \code{data.frame} of cell lines with associated tissue information which is derived in different ways depending on the tissue_info parameter
}

