% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{importCCLE_affy}
\alias{importCCLE_affy}
\title{Import CCLE affy data}
\usage{
importCCLE_affy(fn, con)
}
\arguments{
\item{fn}{The path of the data file}

\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
TRUE or FALSE depending on whether the data has been written successfully
}
\description{
This function imports the information in CCLE_Expression_Entrez_2012-09-29.gct into the ccle_affy table in the database.
Also indexes the table for fast retrieval.
}

