% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importCCLE_cn.R
\name{importCCLE_cn}
\alias{importCCLE_cn}
\title{Import CCLE copy number data}
\usage{
importCCLE_cn(fn, con)
}
\arguments{
\item{fn}{The path of the data file}

\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
TRUE or FALSE depending on whether the data has been written successfully
}
\description{
This function imports the information in CCLE_copynumber_byGene_2012-09-29.txt into the ccle_cn table in the database.
Also indexes the table for fast retrieval.
}

