% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importCCLE_hybcap.R
\name{importCCLE_hybcap}
\alias{importCCLE_hybcap}
\title{Import CCLE hybrid capture sequencing data}
\usage{
importCCLE_hybcap(fn, con)
}
\arguments{
\item{fn}{The path of the data file}

\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
TRUE or FALSE depending on whether the data has been written successfully
}
\description{
This function imports the information in CCLE_hybrid_capture1650_hg19_NoCommonSNPs_NoNeutralVariants_CDS_2012.05.07.xlsx into the ccle_hybap table in the database.
Also indexes the table for fast retrieval.
}

