% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importCellLineIDs.R
\name{importCellLineIDs}
\alias{importCellLineIDs}
\title{Import cell line interconversion data}
\usage{
importCellLineIDs(fn, con)
}
\arguments{
\item{fn}{The path of the data file}

\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
TRUE or FALSE depending on whether the data has been written successfully
}
\description{
This function imports the information in the CellLineIDNormalisationOct15.txt file into the cell_line_ids table in the database.
}

