% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCosmicCLP_exome.R
\name{importCosmicCLP_exome}
\alias{importCosmicCLP_exome}
\title{Import Cosmic Cell Lines Project exome sequencing data}
\usage{
importCosmicCLP_exome(fn, con)
}
\arguments{
\item{fn}{The path of the data file}

\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
TRUE or FALSE depending on whether the data has been written successfully
}
\description{
This function imports the information in CosmicCLP_CompleteExport_v74.tsv into the cosmicclp_exome table in the database.
Also indexes the table for fast retrieval.
}

