% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExportFunctions.R
\name{make_df}
\alias{make_df}
\title{Merge different data types into a single data frame}
\usage{
make_df(con, genes, cell_lines, drugs, data_types = c("affy", "cn", "hybcap",
  "resp", "cosmicclp"), drug_df = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{genes}{A vector of gene symbols}

\item{cell_lines}{A vectore of cell line identifiers}

\item{drugs}{A vector of compound identifiers}

\item{data_types}{A vector with default \code{c('affy', 'hybcap', 'resp')} to specify which data types should be returned.}
}
\value{
A \code{data.frame} containing the affymetrix gene expression data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} containing data for different data types in a form suitable for further statistical modelling in R.
}

