% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VisFunctions.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make a heatmap from a tall data frame}
\usage{
make_heatmap(tall_df, order_feature = NULL)
}
\arguments{
\item{tall_df}{A data frame from make_tall_df}
}
\value{
A ggplot2 object
}
\description{
This function creates a heatmap from the output of the make_tall_df function.
}

