% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShinyVisFunctions.R
\name{make_shiny_df}
\alias{make_shiny_df}
\title{Get data for shiny}
\usage{
make_shiny_df(con, gene, cell_lines, drug, data_types = c("affy", "cn",
  "hybcap", "cosmicclp"), drug_df = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{cell_lines}{A vector of cell line identifiers}

\item{drugs}{A vector of compound identifiers}
}
\value{
A \code{data.frame} containing the drug response data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} suitable for the shiny app
}

