% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ManipulateFunctions.R
\name{make_wide_df}
\alias{make_wide_df}
\title{Turn a tall data frame from make_tall_df into a wide data frame}
\usage{
make_wide_df(df)
}
\arguments{
\item{df}{A \code{data.frame} in tall format generated by \code{make_tall_df}}
}
\value{
A \code{data.frame} in wide format
}
\description{
This function creates a \code{data.frame} containing data for different data types in a form suitable for further statistical modelling in R.
}

