% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeGeneticVsGeneticDataFrame.R
\name{makeGeneticVsGeneticDataFrame}
\alias{makeGeneticVsGeneticDataFrame}
\title{Get GeneticVsGenetic data}
\usage{
makeGeneticVsGeneticDataFrame(con, cell_lines, gene1, gene2, data_type1,
  data_type2)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{cell_lines}{A vector of cell line identifiers}

\item{gene1}{The gene symbol of the primary gene}

\item{gene2}{The gene symbol of the secondary gene}

\item{data_type1}{The data type required for gene1: should be one of \code{'affy', 'cn', 'hybcap', 'cosmicclp'}}

\item{data_type2}{The data type required for gene1: should be one of \code{'affy', 'cn', 'hybcap', 'cosmicclp'}}
}
\value{
A \code{data.frame} containing the combined genetic data for the requested cell lines
}
\description{
This function creates a \code{data.frame} suitable for the GeneticVsGenetic plots and shiny app
}

