% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeRespVsGeneticDataFrame.R
\name{makeRespVsGeneticDataFrame}
\alias{makeRespVsGeneticDataFrame}
\title{Get RespVsGenetic data}
\usage{
makeRespVsGeneticDataFrame(con, gene, cell_lines, drug, data_types = c("affy",
  "cn", "hybcap", "cosmicclp"), drug_df = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{cell_lines}{A vector of cell line identifiers}

\item{drugs}{A vector of compound identifiers}
}
\value{
A \code{data.frame} containing the drug response and genetic data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} suitable for the RespVsGenetic plots and shiny app
}

