% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeRespVsRespDataFrame.R
\name{makeRespVsRespDataFrame}
\alias{makeRespVsRespDataFrame}
\title{Make a RespVsResp data frame}
\usage{
makeRespVsRespDataFrame(con, drugs, cell_lines, resp_type = "ccle",
  tissue_info = "ccle", drug_df = NULL)
}
\arguments{
\item{drugs}{A vector of compound identifiers}

\item{cell_lines}{A vector of cell line identifiers}

\item{resp_type}{The type of response data: \code{ccle, custom}}

\item{tissue_info}{Source of tissue information - see ?getTissueInfo}

\item{drug_df}{A \code{data.frame} containing input for the \code{getDrugData_custom} function}

\item{src}{A \code{SQLiteConnection} object to the database (ccle) or \code{data.frame} object (custom)}
}
\value{
A \code{data.frame} containing the response data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} containing the response data suitable for the RespVsResp plotting functions
}

