% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{makeToyDB}
\alias{makeToyDB}
\title{Set up a toy database}
\usage{
makeToyDB()
}
\arguments{
\item{none}{}
}
\value{
A \code{SQLiteConnection} object
}
\description{
This function imports the example toy dataset included with the package into a temporary database.
}

