% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWideDataFrame.R
\name{makeWideDataFrame}
\alias{makeWideDataFrame}
\title{Merge different data types into a single wide data frame}
\usage{
makeWideDataFrame(con, genes, cell_lines, drugs, data_types = c("affy", "cn",
  "hybcap", "resp", "cosmicclp"), drug_df = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{genes}{A vector of gene symbols}

\item{cell_lines}{A vector of cell line identifiers}

\item{drugs}{A vector of compound identifiers}

\item{data_types}{A vector with default \code{c('affy', 'cn', 'hybcap', 'resp', 'cosmicclp')} to specify which data types should be returned.}

\item{drug_df}{A \code{data.frame} containing input for the \code{getDrugData_custom} function}
}
\value{
A wide \code{data.frame} containing the requested data
}
\description{
This function creates a \code{data.frame} containing data for different data types in a form suitable for further statistical modelling in R.
}

