% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWideFromRespVsRespDataFrame.R
\name{makeWideFromRespVsRespDataFrame}
\alias{makeWideFromRespVsRespDataFrame}
\title{Make a RespVsResp wide data frame}
\usage{
makeWideFromRespVsRespDataFrame(df)
}
\arguments{
\item{df}{A \code{data.frame} object generated by the makeRespVsRespDataFrame function}
}
\value{
A pivotted \code{data.frame} with one response variable per column
}
\description{
This function creates a pivotted \code{data.frame} with the output of the makeRespVsRespDataFrame function
}

