% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeWideFromTallDataFrame.R
\name{makeWideFromTallDataFrame}
\alias{makeWideFromTallDataFrame}
\title{Turn a tall data frame from makeTallDataFrame into a wide data frame}
\usage{
makeWideFromTallDataFrame(df)
}
\arguments{
\item{df}{A \code{data.frame} in tall format generated by \code{makeTallDataFrame}}
}
\value{
A \code{data.frame} in wide format
}
\description{
This function creates a \code{data.frame} containing data for different data types in a form suitable for further statistical modelling in R.
}

