% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotGeneticVsGeneticPoint.R
\name{plotGeneticVsGeneticPoint}
\alias{plotGeneticVsGeneticPoint}
\title{Compare two genetic features (scatter/boxplot)}
\usage{
plotGeneticVsGeneticPoint(df, facet_option = FALSE)
}
\arguments{
\item{df}{A data frame from the \code{makeGeneticVsGeneticDataFrame} function}
}
\value{
A ggplot2 object
}
\description{
Generate a plot that compares two genetic features in a set of cell lines
}

