% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Make a heatmap from a tall data frame}
\usage{
plotHeatmap(tall_df, order_feature = NULL)
}
\arguments{
\item{tall_df}{A data frame from makeTallDataFrame}
}
\value{
A ggplot2 object
}
\description{
This function creates a heatmap from the output of the makeTallDataFrame function.
}

