% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRespVsGeneticHist.R
\name{plotRespVsGeneticHist}
\alias{plotRespVsGeneticHist}
\title{Compare response against a genetic feature (histogram)}
\usage{
plotRespVsGeneticHist(df, data_type = NULL, facet_option = FALSE,
  label_option = FALSE)
}
\arguments{
\item{df}{A data frame from the \code{makeRespVsGeneticDataFrame} function}
}
\value{
A ggplot2 object
}
\description{
Generate a plot that compares response in cell lines with different genetic features
}

