% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRespVsGeneticPoint.R
\name{plotRespVsGeneticPoint}
\alias{plotRespVsGeneticPoint}
\title{Compare response against a genetic feature (scatter/boxplot)}
\usage{
plotRespVsGeneticPoint(df, data_type = NULL, facet_option = FALSE)
}
\arguments{
\item{df}{A data frame from the \code{makeRespVsGeneticDataFrame} function}
}
\value{
A ggplot2 object
}
\description{
Generate a plot that compares response in cell lines with different genetic features
}

