% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRespVsRespPairs.R
\name{plotRespVsRespPairs}
\alias{plotRespVsRespPairs}
\title{Correlation plots for selected response variables}
\usage{
plotRespVsRespPairs(df)
}
\arguments{
\item{df}{A data frame from the \code{makeRespVsRespDataFrame} function}
}
\value{
A ggplot2 object
}
\description{
Generate a set of scatter plots for the selected response variables
}

