% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRespVsRespWaterfall.R
\name{plotRespVsRespWaterfall}
\alias{plotRespVsRespWaterfall}
\title{Waterfall plot for selected response variables}
\usage{
plotRespVsRespWaterfall(df)
}
\arguments{
\item{df}{A data frame from the \code{makeRespVsRespDataFrame} function}
}
\value{
A ggplot2 object
}
\description{
Generate a set of waterfall plots for the selected response variables
}

