% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ShinyVisFunctions.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Add title}
\usage{
run_shiny_app(con, drug_df)
}
\arguments{
\item{drug_df}{A data frame containing the drug data}
}
\value{
returns
}
\description{
add description
}

