% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyGeneticVsGeneticApp.R
\name{shinyGeneticVsGeneticApp}
\alias{shinyGeneticVsGeneticApp}
\title{Genetic vs Genetic Shiny visualisation}
\usage{
shinyGeneticVsGeneticApp(con)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
Launches an interactive Shiny application
}
\description{
Interactive plots from user defined genes and data types in Shiny
}

