% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGeneticVsGeneticServer.R
\name{shinyGeneticVsGeneticServer}
\alias{shinyGeneticVsGeneticServer}
\title{shinyGeneticVsGeneticServer}
\usage{
shinyGeneticVsGeneticServer(input, output, con)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{con}{SQLite connection object}
}
\value{
a shiny server
}
\description{
Create a shiny server for the GeneticsvsGenetic shiny app
}

