% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyRespVsGeneticApp.R
\name{shinyRespVsGeneticApp}
\alias{shinyRespVsGeneticApp}
\title{Runs the Shiny visualisation}
\usage{
shinyRespVsGeneticApp(con, drug_df = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{drug_df}{A data frame containing the drug data}
}
\value{
Launches an interactive Shiny application
}
\description{
Takes a user supplied data frame of drug response data and combines it with user defined gene feature data
}

