% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRespVsGeneticServer.R
\name{shinyRespVsGeneticServer}
\alias{shinyRespVsGeneticServer}
\title{shinyRespVsGeneticServer}
\usage{
shinyRespVsGeneticServer(input, output, con, drug_df = NULL)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{con}{SQLite connection object}

\item{drug_df}{drug data frame}
}
\value{
a shiny server
}
\description{
Create a shiny server for the RespVsGenetic shiny app
}

