% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRespVsRespApp.R
\name{shinyRespVsRespApp}
\alias{shinyRespVsRespApp}
\title{Response vs Response Shiny visualisation}
\usage{
shinyRespVsRespApp(con, drug_df = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}
}
\value{
Launches an interactive Shiny application
}
\description{
Interactive plots comparing and visualising different response variables in Shiny
}

