% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeGeneticVsGeneticDataFrame.R
\name{makeGeneticVsGeneticDataFrame}
\alias{makeGeneticVsGeneticDataFrame}
\title{Get GeneticVsGenetic data}
\usage{
makeGeneticVsGeneticDataFrame(con, cell_lines, gene1, gene2, data_type1,
  data_type2)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{cell_lines}{A vector of cell line identifiers}

\item{drugs}{A vector of compound identifiers}
}
\value{
A \code{data.frame} containing the combined genetic data for the requested cell lines
}
\description{
This function creates a \code{data.frame} suitable for the GeneticVsGenetic plots and shiny app
}

