% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRespVsGeneticDataFrame.R
\name{makeRespVsGeneticDataFrame}
\alias{makeRespVsGeneticDataFrame}
\title{Get RespVsGenetic data}
\usage{
makeRespVsGeneticDataFrame(con, gene, cell_lines, drug, data_types = c("affy",
  "cn", "hybcap", "cosmicclp"), drug_df = NULL, tissue_info = "ccle")
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{gene}{The gene symbol for the gene of interest}

\item{cell_lines}{A vector of cell line identifiers}

\item{drug}{The drug of interest}

\item{data_types}{A vector with default \code{c('affy', 'cn', 'hybcap', 'resp', 'cosmicclp')} to specify which data types should be returned.}

\item{drug_df}{A \code{data.frame} containing input for the \code{getDrugData_custom} function}

\item{tissue_info}{Source of tissue information - see ?getTissueInfo}
}
\value{
A \code{data.frame} containing the drug response and genetic data for the requested compounds and cell lines
}
\description{
This function creates a \code{data.frame} suitable for the RespVsGenetic plots and shiny app
}

