% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeToyDB.R
\name{makeToyDB}
\alias{makeToyDB}
\title{Set up a toy database}
\usage{
makeToyDB()
}
\value{
A \code{SQLiteConnection} object
}
\description{
This function imports the example toy dataset included with the package into a temporary database.
}

