% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneticVsGeneticHist.R
\name{plotGeneticVsGeneticHist}
\alias{plotGeneticVsGeneticHist}
\title{Compare two genetic features (histogram)}
\usage{
plotGeneticVsGeneticHist(df, facet_option = FALSE, label_option = FALSE)
}
\arguments{
\item{df}{A data frame from the \code{makeGeneticVsGeneticDataFrame} function}

\item{facet_option}{Whether or not to facet by tissue (default FALSE)}

\item{label_option}{Whether or not to label cell lines (default FALSE)}
}
\value{
A ggplot2 object
}
\description{
Generate a histogram type plot that compares two genetic features in a set of cell lines
}

