% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupSQLite.R
\name{setupSQLite}
\alias{setupSQLite}
\title{Connect to or create a RSQlite connection}
\usage{
setupSQLite(dbpath = system.file("extdata/toy.db", package =
  "CancerCellLines"))
}
\arguments{
\item{dbpath}{The path of the database}
}
\value{
A \code{SQLiteConnection} object
}
\description{
This function connects to or creates if necessary a SQLite database and returns a SQLiteConnection object.
}

